backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.40.2",
    "pandas>=2.3.1",
    "pandera>=0.25.0",
    "polars>=1.32.0",
    "pyarrow>=21.0.0",
    "pyhive[hive-pure-sasl]>=0.7.0",
    "python-dotenv>=1.1.1",
    "sanic>=25.3.0",
]
